07.06 state
fixing error, not completed

+++ app/src/main/java/com/example/nework/entity/CoordsEmbeddable.kt
@@ -5,8 +5,8 @@ import com.example.nework.dto.Coords
 
 @Entity
 data class CoordsEmbeddable (
-    val lat: Int,
-    val long: Int
+    val lat: Double,
+    val long: Double
 ) {
     fun toDto(): Coords = Coords(
         lat = lat,
